Added map  button.
Now when we click a bar listing and it brings us to the bar page, there is a button which if you hit takes you to whatever mapping app you have and shows you bar location.
Todo:

Cache search results for faster retrieval and to reduce number of Place api calls we do.